preprocessing/FRI/fri_requirements.txt
absl-py==2.1.0
astunparse==1.6.3
biopython==1.76
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
cachetools==5.3.3
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.3.2
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
decorator==5.1.1
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.25.0
gast==0.3.3
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h5py==2.10.0
idna==3.7
importlib-metadata==6.7.0
joblib==1.3.2
keract==4.5.1
Keras-Preprocessing==1.1.2
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib @ file:///croot/matplotlib-suite_1667356714455/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
munkres==1.1.4
networkx==2.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib==3.2.2
opt-einsum==3.3.0
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.3.5
Pillow==9.4.0
ply==3.11
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1671697431263/work
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.23.1
scipy==1.7.3
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
termcolor==2.3.0
threadpoolctl==3.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
